requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
icalendar==4.0.9
idna==3.3
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
git+https://github.com/spralja/django-ical-models@master